db close not required
